Conflicts:
	src/Integration/AsyncGetDataFixture.cs
	src/Integration/MiniMailFixture.cs

+++ src/PrgData/code.asmx.vb
@@ -2972,7 +2972,7 @@ StartZipping:
 			Log.Debug("Попытка запустить ProtocolUpdatesThread.Start()")
 			ProtocolUpdatesThread.Start()
 
-			If UpdateType <> RequestType.ResumeData Then
+			If Not ErrorFlag AndAlso UpdateType <> RequestType.ResumeData Then
 				If File.Exists(UpdateData.GetCurrentFile(GUpdateId)) Then
 					Me.Log.DebugFormat("Производим попытку удаления файла: {0}", UpdateData.GetCurrentFile(GUpdateId))
 					File.Delete(UpdateData.GetCurrentFile(GUpdateId))
@@ -2985,9 +2985,12 @@ StartZipping:
 				connection.ConnectionString = Settings.ConnectionString
 				connection.Open()
 
-				UpdateHelper.UpdateRequestType(connection, UpdateData, GUpdateId, Addition, ResultLenght)
+				If UpdateType = RequestType.Error then
+					AnalitFUpdate.UpdateLog(connection, GUpdateId, UpdateType, Addition)
+				Else 
+					UpdateHelper.UpdateRequestType(connection, UpdateData, GUpdateId, Addition, ResultLenght)
+				End If
 			End Using
-
 
 			Log.Debug("Попытка удалить из списка AsyncPrgDatas.DeleteFromList")
 			AsyncPrgDatas.DeleteFromList(Me)
@@ -4110,6 +4113,7 @@ RestartTrans2:
 			ErrorFlag = True
 			UpdateType = RequestType.Error
 			Addition &= ex.Message
+			PackProtocols()
 		End Try
 	End Sub
